docker_compose/crud/stocks_products/requirements.txt
django
djangorestframework
psycopg2-binary
django-filter
gunicorn
